+++ NAMESPACE
@@ -25,6 +25,7 @@ export(series_to_string)
 export(series_to_wordbag)
 export(str_to_repair_grammar)
 export(subseries)
+export(ts_to_intervals)
 export(znorm)
 importFrom(Rcpp,sourceCpp)
 useDynLib(jmotif)
