+++ R/RcppExports.R
@@ -86,9 +86,6 @@ paa <- function(ts, paa_num) {
 }
 
 #' Runs the repair on a string.
-NULL
-
-#' Runs the repair on a string.
 #'
 #' @param str the input string.
 #' @useDynLib jmotif
@@ -99,6 +96,12 @@ NULL
 #' str_to_repair_grammar("abc abc cba cba bac xxx abc abc cba cba bac")
 str_to_repair_grammar <- function(str) {
     .Call('jmotif_str_to_repair_grammar', PACKAGE = 'jmotif', str)
+}
+
+#' @useDynLib jmotif
+#' @export
+ts_to_intervals <- function(ts, w_size, paa_size, a_size, nr_strategy, n_threshold = 0.01) {
+    .Call('jmotif_ts_to_intervals', PACKAGE = 'jmotif', ts, w_size, paa_size, a_size, nr_strategy, n_threshold)
 }
 
 #' Translates an alphabet size into the array of corresponding SAX cut-lines built using the Normal distribution.
