+++ inst/include/jmotif.h
@@ -98,13 +98,90 @@ Rcpp::DataFrame find_discords_hot_sax(NumericVector ts, int w_size, int paa_size
                                       int a_size, double n_threshold, int discords_num);
 
 //
+// REPAIR
+//
+
+// Tokens are used in the R0
+//
+class Token {
+public:
+  int str_idx;
+  std::string payload;
+  Token(){
+    str_idx = -1;
+  };
+  Token(std::string s, int idx){
+    payload = s;
+    str_idx = idx;
+  };
+//   std::ostream& operator<<(std::ostream &strm, const Token &t) {
+//     return strm << "T(" << t.payload << " @ " << t.str_idx << ")";
+//   };
+};
+
+// Rules build up the rules table, i.e. the grammar
+//
+class Rule {
+public:
+  int id;
+  std::string rule_string;
+  std::string expanded_rule_string;
+  std::vector<int> occurrences;
+  Rule(){
+    id = -1;
+    rule_string = "\0";
+    expanded_rule_string = "\0";
+  };
+  Rule(int r_id, std::string rule_str, std::string expanded_rule_str){
+    id = r_id;
+    rule_string = rule_str;
+    expanded_rule_string = expanded_rule_str;
+  };
+  std::string ruleString(){
+    std::stringstream ss;
+    ss << id;
+    return "R" + ss.str();
+  };
+//   std::ostream& operator<<(std::ostream &strm, const Rule &d) {
+//     return strm << "R" << d.id << "\t" << d.rule_string << "\t" << d.expanded_rule_string;
+//   };
+};
+
+
+// Guards are the placeholders for tokens
+//
+class Guard: public Token {
+public:
+  Rule r;
+  Guard(Rule rule, int idx){
+    r = rule;
+    payload = r.ruleString();
+    str_idx = idx;
+  };
+};
+
+// Rule records make the REPAIR output
+//
+class RuleRecord {
+public:
+  int rule_id;
+  std::string rule_string;
+  std::string expanded_rule_string;
+  std::vector<int> rule_occurrences;
+  std::vector<std::pair<int,int>> rule_intervals;
+};
+
+Rcpp::List str_to_repair_grammar(CharacterVector str);
+
+//
 // Utilities
 //
 NumericVector col_means(NumericMatrix m);
 //
 NumericVector subseries(NumericVector ts, int start, int end);
+//
+int count_spaces(std::string *s);
 
-Rcpp::List str_to_repair_grammar(CharacterVector str);
 
 // internal high performance computing
 //
