+++ src/RcppExports.cpp
@@ -84,6 +84,22 @@ BEGIN_RCPP
     return __result;
 END_RCPP
 }
+// ts_to_intervals
+Rcpp::NumericVector ts_to_intervals(NumericVector ts, int w_size, int paa_size, int a_size, CharacterVector nr_strategy, double n_threshold);
+RcppExport SEXP jmotif_ts_to_intervals(SEXP tsSEXP, SEXP w_sizeSEXP, SEXP paa_sizeSEXP, SEXP a_sizeSEXP, SEXP nr_strategySEXP, SEXP n_thresholdSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject __result;
+    Rcpp::RNGScope __rngScope;
+    Rcpp::traits::input_parameter< NumericVector >::type ts(tsSEXP);
+    Rcpp::traits::input_parameter< int >::type w_size(w_sizeSEXP);
+    Rcpp::traits::input_parameter< int >::type paa_size(paa_sizeSEXP);
+    Rcpp::traits::input_parameter< int >::type a_size(a_sizeSEXP);
+    Rcpp::traits::input_parameter< CharacterVector >::type nr_strategy(nr_strategySEXP);
+    Rcpp::traits::input_parameter< double >::type n_threshold(n_thresholdSEXP);
+    __result = Rcpp::wrap(ts_to_intervals(ts, w_size, paa_size, a_size, nr_strategy, n_threshold));
+    return __result;
+END_RCPP
+}
 // alphabet_to_cuts
 NumericVector alphabet_to_cuts(int a_size);
 RcppExport SEXP jmotif_alphabet_to_cuts(SEXP a_sizeSEXP) {
