+++ src/utils.cpp
@@ -43,3 +43,11 @@ NumericVector subseries(NumericVector ts, int start, int end) {
   }
   return res;
 }
+
+int count_spaces(std::string *s) {
+  int count = 0;
+  for (int i = 0; i < s->size(); i++)
+    if (s->at(i) == ' ') count++;
+    return count;
+}
+
